+++ chains/substrate/config.go
@@ -36,7 +36,7 @@ type SubstrateConfig struct {
 }
 
 func (c *SubstrateConfig) String() string {
-	kp, _ := signature.KeyringPairFromSecret(c.GeneralChainConfig.Key, c.SubstrateNetwork)
+	kp, _ := signature.KeyringPairFromSecret(c.GeneralChainConfig.Key, uint16(c.SubstrateNetwork))
 	return fmt.Sprintf(`Name: '%s', Id: '%d', Type: '%s', BlockstorePath: '%s', FreshStart: '%t', 
 							  LatestBlock: '%t', Key address: '%s', StartBlock: '%s', BlockInterval: '%s', 
                               BlockRetryInterval: '%s', ChainID: '%d', Tip: '%d', SubstrateNetworkPrefix: "%d"`,
